PO-2255 - tidy and add tests

+++ src/main/java/uk/gov/hmcts/opal/service/report/OperationReportByEnforcementTransaction.java
@@ -11,8 +11,8 @@ public class OperationReportByEnforcementTransaction implements ReportDataInterf
     private ReportMetaData reportMetaData;
 
     @Override
-    public long getNumberOfRecords() {
-        return transactionList == null ? 0 : transactionList.size();
+    public short getNumberOfRecords() {
+        return transactionList == null ? (short) 0 : (short) transactionList.size();
     }
 
     @Override
